refactor object attributes
